新增 登录界面（已基本完成，还差几个图标）

+++ FunctionClass/JHRefresh/JHRefreshConfig.m
@@ -43,10 +43,9 @@ NSString *const JHRefreshLastUpdateTimeFormat = @"yyyy-MM-dd HH:mm";
 {
     NSUserDefaults *userDefault = [NSUserDefaults standardUserDefaults];
     NSDictionary *refreshDic = [userDefault objectForKey:JHRefreshConfigKey];
-    NSString *time = [[refreshDic objectForKey:JHRefreshLastUpdateTimeKey] objectForKey:[NSString stringWithFormat:@"%@_%d",JHRefreshLastUpdateTimeKey,ID]];
+    NSString *time = [[refreshDic objectForKey:JHRefreshLastUpdateTimeKey] objectForKey:[NSString stringWithFormat:@"%@_%ld",JHRefreshLastUpdateTimeKey,(long)ID]];
     
-    if(!time)
-    {
+    if(!time) {
         return [[self class] nowTimeString];
     }
     
@@ -66,7 +65,7 @@ NSString *const JHRefreshLastUpdateTimeFormat = @"yyyy-MM-dd HH:mm";
     {
         lastUpdateTimeDic = [NSMutableDictionary dictionary];
     }
-    [lastUpdateTimeDic setObject:[[self class] nowTimeString] forKey:[NSString stringWithFormat:@"%@_%d",JHRefreshLastUpdateTimeKey,ID]];
+    [lastUpdateTimeDic setObject:[[self class] nowTimeString] forKey:[NSString stringWithFormat:@"%@_%ld",JHRefreshLastUpdateTimeKey,(long)ID]];
     
     [userDefault setObject:refreshDic forKey:JHRefreshConfigKey];
     [userDefault synchronize];
